pyproject.toml
[project]
name = "btp-ii-lfgd"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26.0",
    "scipy>=1.13.0",
    "datasets>=2.19.0",
    "sentence-transformers>=2.7.0",
    "rank-bm25>=0.2.2",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
    "python-dotenv>=1.0.0",
    "openai>=1.30.0",
    "groq>=0.9.0",
    "vaderSentiment>=3.3.2",
    "modal>=1.4.1",
    "qdrant-client>=1.17.1",
]

[dependency-groups]
dev = ["pytest>=8.0.0"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["lfgd*", "metrics*", "baselines*", "generation*", "retrieval*", "data*", "experiments*", "analysis*"]

# Modal GPU packages - install via Modal app image
[project.optional-dependencies]
modal = [
    "modal>=0.67.0",
    "torch>=2.0.0",
    "transformers>=4.40.0",
]
